Separated the photo size menu from the component with the photo itself
